requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
bs4==0.0.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.6.0
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-browser-reload==1.13.0
django-tailwind==3.8.0
djangorestframework==3.15.2
h11==0.14.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jwt==1.3.1
kombu==5.3.7
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nav==5.3.1
outcome==1.3.0.post0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycparser==2.22
Pygments==2.18.0
PyJWT==2.1.0
PySocks==1.7.1
pyspnego==0.11.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
redis==5.0.7
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rich==13.7.1
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stripe==10.2.0
text-unidecode==1.3
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zeep==4.2.1
